Extrae clases de GameState a archivos separados
